### 2021-09-30 #85 @whedon
[PRE REVIEW]: Computing Reachable Sets of Semi-Discrete Solid Dynamics Equations with ReachabilityAnalysis.jl
**Submitting author:** @mforets (<a href="http://orcid.org/0000-0002-9831-7801">Marcelo Forets</a>)
**Repository:** <a href="https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21" target ="_blank">https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21</a>
**Version:**
