### 2021-10-17 JuliaReach/SetPropagation-FEM-JuliaCon21#1 @PetrKryslUCSD
FEM sw linkages
It seems to me that given this is a contribution to JuliaCon a link to a Julia FEM package should be a priority. Instead of Matlab/Octave...
