### 2021-10-18 #85 @carstenbauer
Well, the JuliaCon proceedings pendant of your JOSS link is https://github.com/JuliaCon/proceedings-review/issues. This thread is the Pre-review thread. Once I've found a second reviewer, we can officially move to review stage (this issue will get closed and a new one with `[REVIEW]` will get opened.
