+++ README.md
@@ -48,7 +48,7 @@ https://github.com/JuliaReach/SetPropagation-FEM-Examples
 
 [4] Luca Feranti, Marcelo Forets, and David P. Sanders. IntervalLinearAlgebra.jl: linear algebra done rigorously, 9 2021. doi:10.5281/zenodo.5363563.
 
-[5] Marcelo Forets, Daniel Freire Caporale, and Jorge M Pérez Zerpa. Combining set propagation with finite element methods for time integration in transient solid mechanics problems. arXiv preprint arXiv:2105.05841, 2021.
+[5] Marcelo Forets, Daniel Freire Caporale, and Jorge M Pérez Zerpa. Combining set propagation with finite element methods for time integration in transient solid mechanics problems. arXiv preprint arXiv:2105.05841. Accepted in Computers & Structures, 2021.
 
 [6] Marcelo Forets, Christian Schilling, and Luca Feranti. IntervalMatrices.jl: Matrices with interval coefficients in Julia, 9 2021.
 doi:10.5281/zenodo.5516249.
