+++ paper/ref.bib
@@ -34,7 +34,7 @@
 @article{forets2021combining,
 	title={Combining Set Propagation with Finite Element Methods for Time Integration in Transient Solid Mechanics Problems},
 	author={Forets, Marcelo and Caporale, Daniel Freire and Zerpa, Jorge M P{\'e}rez},
-	journal={arXiv preprint arXiv:2105.05841},
+	journal={arXiv preprint arXiv:2105.05841. Accepted in Computers \& Structures},
 	year={2021}
 }
 
