+++ paper/ref.bib
@@ -95,3 +95,35 @@
 	title = {{Finite Element Procedures}},
 	year = {2014}
 }
+
+@article{Gridap,
+	doi = {10.21105/joss.02520},
+	url = {https://doi.org/10.21105/joss.02520},
+	year = {2020},
+	publisher = {The Open Journal},
+	volume = {5},
+	number = {52},
+	pages = {2520},
+	author = {Santiago Badia and Francesc Verdugo},
+	title = {Gridap: An extensible Finite Element toolbox in Julia},
+	journal = {Journal of Open Source Software}
+}
+
+@software{Ferrite,
+	author = {Carlsson, Kristoffer and Ekre, Fredrik and Contributors,},
+	month = {3},
+	title = {{Ferrite.jl}},
+	url = {https://github.com/Ferrite-FEM/Ferrite.jl},
+	version = {0.3.0},
+	year = {2021}
+}
+
+@software{FinEtools,
+	author = {Petr Krysl and Contributors,},
+	month = {10},
+	title = {{FinEtools.jl}},
+	url = {https://github.com/PetrKryslUCSD/FinEtools.jl},
+	version = {0.5.6},
+	year = {2021}
+}
+
