+++ paper/paper.tex
@@ -118,7 +118,7 @@ solB = solve(prob, 50, LGG09(δ=5e-2, dirs=:oct, dim=5))
 %
 Probabilistic reachability, and modeling nonlinear behaviors using state-space abstraction methods, are also planned.
 %
-Julia has a thriving ecosystem of open source FEM projects~\cite{Gridap,Ferrite,FinEtools}, and we think that integrating set propagation into some of those tools is the next key step for solving real world problems using reachability.
+Julia has a thriving ecosystem of open source FEM projects~\cite{Gridap,Ferrite,FinEtools}, and we think that integrating set propagation into such tools is the next key step for solving real world problems using reachability.
 
 \input{bib.tex}
 	
