+++ paper/paper.tex
@@ -62,9 +62,9 @@ a system of second-order differential equations in time, where $\mathbf{x}(0) \i
 	
 \vspace{0.2cm}
 
-\emph{Representing solutions with sets.} In \cite{forets2021combining} we present a novel approach for time integration of solid heat transfer and structural dynamics equations based on reachability analysis techniques~\cite{althoff2020set}. These methods are implemented in \href{http://github.com/JuliaReach/ReachabilityAnalysis.jl}{ReachabilityAnalysis.jl}, a core package of \href{http://juliareach.com}{JuliaReach}, while FEM matrices assembly is done using the Octave tool~\href{http://onsas.org}{ONSAS}~\cite{onsas}.
+\emph{Representing solutions with sets.} In \cite{forets2021combining} we present a novel approach for time integration of heat transfer and structural dynamics equations based on reachability analysis techniques~\cite{althoff2020set}. These methods are implemented in \href{http://github.com/JuliaReach/ReachabilityAnalysis.jl}{ReachabilityAnalysis.jl}, a core package of \href{http://juliareach.com}{JuliaReach}, while FEM matrices assembly is done using the Octave tool~\href{http://onsas.org}{ONSAS}~\cite{onsas}.
 %
-Reachability is a modern computational approach where solutions to differential equations are represented using sets.
+Reachability is a modern computational approach where differential equations solutions are represented using sets.
 %
 The set-based conservative time integration of Eq.~\eqref{eq:second_order} returns solution sets (\emph{flowpipes}) that include all exact trajectories, with convergence to the true reachable states as the time step decreases.
 
