### 2021-10-21 JuliaReach/SetPropagation-FEM-JuliaCon21#6 @jorgepz
Bib error
https://github.com/JuliaCon/proceedings-review/issues/95#issuecomment-948724061
