+++ paper/ref.bib
@@ -84,8 +84,8 @@
 	pages     = {250 - 262},
 	year      = {2010},
 	note      = {IFAC World Congress 2008},
-	@url       = {https://@doi.org/10.1016/j.nahs.2009.03.002},
-	@doi       = {10.1016/j.nahs.2009.03.002}
+	url       = {https://@doi.org/10.1016/j.nahs.2009.03.002},
+	doi       = {10.1016/j.nahs.2009.03.002}
 }
 
 @book{Bathe2014,
