+++ paper/ref.bib
@@ -2,7 +2,7 @@
 	author = {
 	Marcelo Forets and
 	Christian Schilling and
-	Luca Feranti 
+	Luca Feranti
 	},
 	title  = {IntervalMatrices.jl: Matrices with interval coefficients in Julia},
 	month  = {9},
@@ -46,7 +46,8 @@
 	number={2},
 	pages={250--262},
 	year={2010},
-	publisher={Elsevier}
+	publisher={Elsevier},
+	doi={10.1016/j.nahs.2009.03.002}
 }
 
 @article{althoff2020set,
@@ -55,7 +56,8 @@
 	journal={Annual Review of Control, Robotics, and Autonomous Systems},
 	volume={4},
 	year={2020},
-	publisher={Annual Reviews}
+	publisher={Annual Reviews},
+	doi={10.1146/annurev-control-071420-081941}
 }
 
 @misc{onsas,
@@ -71,7 +73,8 @@
 	title        = {{J}ulia{R}each},
 	month        = may,
 	year         = 2021,
-	howpublished = {\url{https://github.com/JuliaReach/}}
+	howpublished = {\url{https://github.com/JuliaReach/}},
+	doi = {10.1145/3302504.3311804}
 }
 
 @article{LeGuernic2010250,
@@ -126,4 +129,3 @@
 	version = {0.5.6},
 	year = {2021}
 }
-
