### 2021-10-31 JuliaReach/SetPropagation-FEM-JuliaCon21#12 @PetrKryslUCSD
The citations to repos need URLs
Ferrite, FinEtools citations do not have a url in the references.
