### 2021-11-02 3a961f3
Update paper/paper.tex
Co-authored-by: Jorge Pérez Zerpa <42485529+jorgepz@users.noreply.github.com>
