+++ paper/paper.tex
@@ -104,7 +104,7 @@ solB = solve(prob, 50, LGG09(δ=5e-2, dirs=:oct, dim=5))
 	
 	\includegraphics[width=0.8\linewidth,keepaspectratio]{example/displacement_vs_displacement}
 	%
-	\caption{Flowpipe using canonical directions projected on time (top), and using octagonal directions at displacement coordinates (bottom). We additionally plot random simulations.}
+	\caption{Flowpipe of displacements of mass $2m$ projected on time computed using canonical directions  (top), and flowpipe of displacements computed using octagonal directions (bottom). We additionally plot random simulations.}
 	\label{fig:example}
 \end{figure}
 
