+++ README.md
@@ -32,9 +32,9 @@ The initial-value problem is instantiated and homogeneized as described in [1].
 
 To illustrate the flexibility of our approach, two algorithm choices are considered, both relying on support functions [7] (`LGG09` algorithm). `solA` contains the flowpipe efficiently computed along box directions. To improve the accuracy, `solB` is computed using octagonal template directions.
 
-<img src="https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/blob/main/paper/example/displacement_vs_displacement.png" width="400"/>
+<img src="https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/blob/main/paper/example/displacement_vs_time.png" width="400"/>
 
-<img src="https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/blob/main/paper/example/velocity_vs_displacement.png" width="400"/>
+<img src="https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/blob/main/paper/example/displacement_vs_displacement.png" width="400"/>
 
 ## Examples repository
 
