### 2021-11-09 JuliaReach/SetPropagation-FEM-JuliaCon21#15 @carstenbauer
The corresponding JuliaCon proceedings REVIEW thread containing the current article proof (PDF) is here: https://github.com/JuliaCon/proceedings-review/issues/95
