### 2021-11-09 JuliaReach/SetPropagation-FEM-JuliaCon21#15 @mforets
> there is a figure with a broken link as well.

fixed, thanks.

> The corresponding JuliaCon proceedings REVIEW thread containing the current article proof (PDF) is here:

ok thanks @carstenbauer -- then perhaps adding the `paper.pdf` to this repo is not a good idea (asking since its the first time we submit to a JOSS like system)
