### 2021-11-09 #95 @mforets
Good catch!

Line 6 should be instead `SecondOrderAffineContinuousSystem`. The code to reproduce the pictures is [here](https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/blob/main/paper/example/example.jl).
