### 2021-11-26 #95 @lmiq
Dear authors, my impression about the resume is the following: it is not clear to me how strict are the limitations for the length of the paper, but it seems that you have still space for an additional example. The minimal example provided is too minimal in my opinion. The system of the damped oscillator can be trivially modelled without any package and solved with some integration algorithm from DifferentialEquations, thus it is unclear to me (not being of the field) what facility the package advertised brings to the table. The current resume seems just a link to the actual paper and the package repos, and might be useful for an audience that understands the methods and possible difficulties in modelling the true systems for which the package was developed (thus, people that is already searching for a package that does what is being presented).

In my opinion it would be interesting, if the goal if to call the attention of people that is not too much into the field, to provide one example of the modelling of something more complex, that at least is not trivial to model without any packages. It does not necessarily needs to be a complete (with code) example, but the reference of an actual more interesting application, with connections to the modelling strategy provided by the package made clear.
