### 2021-11-26 #95 @mforets
Dear @lmiq,

Thanks a lot for your feedback!

> it is not clear to me how strict are the limitations for the length of the paper, but it seems that you have still space for an additional example.

I like the idea of having another example, although I wouldn't change the current one (see next comment).

@carstenbauer  do you have any comment about how strict is the 1 page limit, given this reviewers' suggestion?

> The minimal example provided is too minimal in my opinion

We chose a minimal example on purpose, so that Julia users can see and play with a model that can be defined quite easily in a few LOC. That the model itself is studied in undergrad courses I think is a plus (vs defining a model that requires further background etc).

> The system of the damped oscillator can be trivially modelled without any package and solved with some integration algorithm from DifferentialEquations

No, DifferentialEquations and numerical ODE solvers in general model and solve a different mathematical problem!

In the set-based approach one propagates a *set* through an ODE or some other system DAEs, NNs,.. It is not an initial point, not an ensemble of points, it is the whole set, which can be bounded, unbounded, convex or not, etc. That is why we need a lot of infrastructure to *model* set-based problems using specialized libraries that scale and have expressive power, in our case it is [LazySets.jl](https://github.com/JuliaReach/LazySets.jl). Then, we use such libraries to write set-based algorithms (as in ReachabilityAnalysis.jl). See also [here](https://juliareach.github.io/ReachabilityAnalysis.jl/dev/man/faq/#Are-there-other-tools-that-perform-reachability-analysis?) for other related tools, and the [LazySets article](https://raw.githubusercontent.com/JuliaCon/proceedings-papers/jcon.00097/jcon.00097/10.21105.jcon.00097.pdf) for additional context.

>  what facility the package advertised brings to the table

Thanks for the question. It matters to us that it is being addressed properly in the current submission, will look into it, although for space constraints, we can only hope for some general comments and references to the relevant literature. Please note that more "challenging" problems can be found on the related published article: https://www.sciencedirect.com/science/article/abs/pii/S0045794921002212  and accompanying Julia code. A repository of reproducible examples related to the FEM application is being worked on in https://github.com/JuliaReach/SetPropagation-FEM-Examples/

> The current resume seems just a link to the actual paper and the package repos, and might be useful for an audience that understands the methods and possible difficulties in modelling the true systems for which the package was developed (thus, people that is already searching for a package that does what is being presented).

> In my opinion it would be interesting, if the goal if to call the attention of people that is not too much into the field, to provide one example of the modelling of something more complex,

Thanks for these comments. We shall review the current status of this submission accordingly.  The main target audience of this submission are Julia scientific developers in general.
