### 2021-11-26 #95 @lmiq
Nice, the scope of the package is clearer to me now. Maybe if you can be more explicit (or perhaps didactic is the word), like you did here, in the resume, that would attract the attention of other people besides the ones that already have a clear picture of the problem. The accompanying articles are very interesting, I am not completely sure how self-contained the resume presented here is expected to be.

> In the set-based approach one propagates a set through an ODE or some other system DAEs, NNs,.. It is not an initial point, not an ensemble of points, it is the whole set, which can be bounded, unbounded, convex or not, etc. That is why we need a lot of infrastructure to model set-based problems using specialized libraries that scale and have expressive power

If the resume can start with something more with this tone, that would certainly had attracted my attention much more. For a programmer from another field, this specifies a problem which I clearly have no idea how to model, and makes me interested in how these things are implemented, independently of the specific application.
