+++ README.md
@@ -11,7 +11,7 @@ by [Jorge M. Pérez Zerpa](https://github.com/jorgepz) (Instituto de Estructuras
 
 ## Extended Abstract
 
-[paper.pdf](https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/blob/main/paper/paper.pdf)
+During the review process, the pdf is automatically generated by `whedon`, see: https://github.com/JuliaCon/proceedings-review/issues/95
 
 ## Pre-print
 
