+++ README.md
@@ -13,9 +13,11 @@ by [Jorge M. Pérez Zerpa](https://github.com/jorgepz) (Instituto de Estructuras
 
 During the review process, the pdf is automatically generated by `whedon`, see: https://github.com/JuliaCon/proceedings-review/issues/95
 
-## Pre-print
+## Publication
 
-https://arxiv.org/abs/2105.05841
+See Forets, Marcelo, Daniel Freire Caporale, and Jorge M. Pérez Zerpa. ["Combining set propagation with finite element methods for time integration in transient solid mechanics problems."](https://www.sciencedirect.com/science/article/abs/pii/S0045794921002212) Computers & Structures 259 (2022): 106699.
+
+The pre-print is available here: https://arxiv.org/abs/2105.05841
 
 ## Minimal example
 
