### 2022-01-21 JuliaReach/SetPropagation-FEM-JuliaCon21#15 @mforets
> I would wait with adding it until the review process is finished.

Thanks for the clarification, we've updated the README correspondingly.
