+++ paper/paper.tex
@@ -52,39 +52,54 @@
 
 \maketitle
 
-\emph{From continuum to semi-discrete equations.} In the context of linear solid dynamics, the spatial discretization of the governing partial differential equations (PDEs) using the Finite-Element Method (FEM)~\cite{Bathe2014}, results in
+\emph{The set-based approach.} %
+%
+Many real-world problems require the resolution of ODEs with uncertainties such as: initial conditions or applied loads. %
+%
+Obtaining solutions considering these uncertainties is a challenging task, particularly in large scale systems.
+%
+The set-based approach consists in the construction of sets that contain all the feasible solutions of the ODEs. %
+% ---------------
+
+
+\vspace{0.2cm}
+
+\emph{Solid Dynamics ODEs.} %
+%
+In problems such as wave propagation or structural vibrations, solid dynamics problems are formulated. %
+%
+In these cases, large systems of ODEs of the form:
 %
 \begin{equation}\label{eq:second_order}
 \mathbf{M} \mathbf{x}''(t) + \mathbf{C}\mathbf{x}'(t) + \mathbf{K}\mathbf{x}(t) = \mathbf{F}(t),
 \end{equation}
 %
-a system of second-order differential equations in time, where $\mathbf{x}(0) \in \mathcal{X}_0$ and $\mathbf{x}'(0) \in \mathcal{V}_0$ are the sets of initial displacements and velocities, $\mathbf{x} \in \mathbb{R}^n$ is the state vector, and $\mathbf{M}$, $\mathbf{C}$ and $\mathbf{K}$ are the mass, damping and stiffness matrices, respectively. Depending on the problem and the mesh size, $n$ is typically  between $10^2$ and $10^5$.
+are formulated using the Finite-Element Method (FEM)~\cite{Bathe2014}, where $\mathbf{x} \in \mathbb{R}^n$ is the displacements (or state) vector, and $\mathbf{M}$, $\mathbf{C}$ and $\mathbf{K}$ are the mass, damping and stiffness matrices, respectively. Depending on the problem $n$ is typically  between $10^2$ and $10^5$.
+
 	
 \vspace{0.2cm}
 
-\emph{Representing solutions with sets.} In \cite{forets2021combining} we present a novel approach for time integration of heat transfer and structural dynamics equations based on reachability analysis techniques~\cite{althoff2020set}. These methods are implemented in \href{http://github.com/JuliaReach/ReachabilityAnalysis.jl}{ReachabilityAnalysis.jl}, a core package of \href{http://juliareach.com}{JuliaReach}, while FEM matrices assembly is done using the Octave tool~\href{http://onsas.org}{ONSAS}~\cite{onsas}.
+\emph{Set-based Solid dynamics.} %
 %
-Reachability is a modern computational approach where differential equations solutions are represented using sets.
+When uncertainty is present in the initial conditions, the displacements and velocities can be considered in sets as: $\mathbf{x}(0) \in \mathcal{X}_0$ and $\mathbf{x}'(0) \in \mathcal{V}_0$. %
 %
-The set-based conservative time integration of Eq.~\eqref{eq:second_order} returns solution sets (\emph{flowpipes}) that include all exact trajectories, with convergence to the true reachable states as the time step decreases.
+In \cite{forets2021combining} a novel approach for time integration of structural dynamics equations based on set-based techniques~\cite{althoff2020set} is presented. %
+%
+This approach allows to compute in a single integration the solution sets (or \emph{flowpipes}) that include all exact trajectories for uncertainties in initial conditions and applied loads. %
+%
 
-\vspace{-0.1cm}
 
[…]
