### 2022-01-21 JuliaReach/SetPropagation-FEM-JuliaCon21#16 @mforets
Thanks for your input and sorry for the very late reply.

We have added a Manifest.toml file to the examples repo (https://github.com/JuliaReach/SetPropagation-FEM-Examples#how-to-run-this-package). We have tried it on a Windows machine running Julia v1.7 and it worked.
