+++ paper/paper.tex
@@ -94,7 +94,7 @@ This approach allows to compute in a single integration the solution sets (or \e
 
 \emph{Set-based FEM implementation and application.} %
 %
-In order to apply the set-based approach for solid dynamics, the package \href{http://github.com/JuliaReach/ReachabilityAnalysis.jl}{ReachabilityAnalysis.jl}, was extended, providing a simple interface that can also be connected with FEM analysis tools such as \href{http://onsas.org}{ONSAS}~\cite{onsas}.
+In order to apply the set-based approach for solid dynamics, the package \href{http://github.com/JuliaReach/ReachabilityAnalysis.jl}{ReachabilityAnalysis.jl}\cite{ReachabilityAnalysis} was extended, providing a simple interface that can also be connected with FEM analysis tools such as \href{http://onsas.org}{ONSAS}~\cite{onsas}.
 %
 The implementations developed can be used to solve large systems, however, in this abstract a minimal didactic problem is solved. %
 %
