+++ paper/paper.tex
@@ -52,13 +52,13 @@
 
 \maketitle
 
-\emph{The set-based approach.} %
+\emph{The Set-Based Approach.} %
 %
-Many real-world problems require the resolution of ODEs with uncertainties such as: initial conditions or applied loads. %
+Many real-world problems require the resolution of ODEs with uncertainties such as initial conditions or applied loads. %
 %
 Obtaining solutions considering these uncertainties is a challenging task, particularly in large scale systems.
 %
-The set-based approach consists in the construction of sets that contain all the feasible solutions of the ODEs. %
+The set-based approach consists in the construction of sets that contain all the feasible solutions of the ODEs~\cite{althoff2020set}. %
 % ---------------
 
 
@@ -79,74 +79,76 @@ are formulated using the Finite-Element Method (FEM)~\cite{Bathe2014}, where $\m
 	
 \vspace{0.2cm}
 
-\emph{Set-based Solid dynamics.} %
+\emph{Set-Based Solid Dynamics.} %
 %
-When uncertainty is present in the initial conditions, the displacements and velocities can be considered in sets as: $\mathbf{x}(0) \in \mathcal{X}_0$ and $\mathbf{x}'(0) \in \mathcal{V}_0$. %
+When uncertainty is present, the initial displacements and velocities are modeled as $\mathbf{x}(0) \in \mathcal{X}_0$ and $\mathbf{x}'(0) \in \mathcal{V}_0$ respectively. %
 %
-In \cite{forets2021combining} a novel approach for time integration of structural dynamics equations based on set-based techniques~\cite{althoff2020set} is presented. %
+In \cite{forets2021combining} a novel approach for time integration of structural dynamics equations based on set-based techniques was presented. It allows to compute, in a single integration, the solution sets (or \emph{flowpipes}) that include all exact trajectories under uncertainties in the initial conditions and applied loads.
 %
-This approach allows to compute in a single integration the solution sets (or \emph{flowpipes}) that include all exact trajectories for uncertainties in initial conditions and applied loads. %
-%
-
+Such solution sets cannot be obtained with standard numerical integrators because they are designed to propagate initial points, not sets.
 
 \vspace{0.2cm}
 
 
-\emph{Set-based FEM implementation and application.} %
+\emph{Set-Based FEM Implementation and Application.} %
 %
-In order to apply the set-based approach for solid dynamics, the package \href{http://github.com/JuliaReach/ReachabilityAnalysis.jl}{ReachabilityAnalysis.jl}\cite{ReachabilityAnalysis} was extended, providing a simple interface that can also be connected with FEM analysis tools such as \href{http://onsas.org}{ONSAS}~\cite{onsas}.
+We have extended the package \href{http://github.com/JuliaReach/ReachabilityAnalysis.jl}{ReachabilityAnalysis.jl}\cite{ReachabilityAnalysis} to support the set-based approach for solid dynamics. Through a simple interface, it can be integrated with FEM analysis tools such as \href{http://onsas.org}{ONSAS}~\cite{onsas}.
 %
-The implementations developed can be used to solve large systems, however, in this abstract a minimal didactic problem is solved. %
+The implementations developed can be used to solve large systems, however, in this abstract a minimal illustrative problem was chosen.
 %
[…]
