+++ paper/ref.bib
@@ -36,7 +36,7 @@
 
 @article{forets2021combining,
 	title = {Combining set propagation with finite element methods for time integration in transient solid mechanics problems},
-	journal = {Computers & Structures},
+	journal = {Computers \& Structures},
 	volume = {259},
 	pages = {106699},
 	year = {2022},
