+++ paper/paper.tex
@@ -54,9 +54,9 @@
 
 \emph{The Set-Based Approach.} %
 %
-Many real-world problems require the resolution of ODEs with uncertainties such as initial conditions or applied loads. %
+Real-world problems usually require the resolution of ODEs with uncertainties in the initial conditions or in the constant term. %
 %
-Obtaining solutions considering these uncertainties is a challenging task, particularly in large scale systems.
+Obtaining solutions considering these uncertainties is a challenging task, particularly in large scale systems. %
 %
 The set-based approach consists in the construction of sets that contain all the feasible solutions of the ODEs~\cite{althoff2020set}. %
 % ---------------
@@ -68,13 +68,13 @@ The set-based approach consists in the construction of sets that contain all the
 %
 In problems such as wave propagation or structural vibrations, solid dynamics problems are formulated. %
 %
-In these cases, large systems of ODEs of the form:
+In these cases, large ODEs systems of the form:
 %
 \begin{equation}\label{eq:second_order}
 \mathbf{M} \mathbf{x}''(t) + \mathbf{C}\mathbf{x}'(t) + \mathbf{K}\mathbf{x}(t) = \mathbf{F}(t),
 \end{equation}
 %
-are formulated using the Finite-Element Method (FEM)~\cite{Bathe2014}, where $\mathbf{x} \in \mathbb{R}^n$ is the displacements (or state) vector, and $\mathbf{M}$, $\mathbf{C}$ and $\mathbf{K}$ are the mass, damping and stiffness matrices, respectively. Depending on the problem $n$ is typically  between $10^2$ and $10^5$.
+are formulated using the Finite-Element Method (FEM)~\cite{Bathe2014}, where $\mathbf{x} \in \mathbb{R}^n$ is the displacements (or state) vector, and $\mathbf{M}$, $\mathbf{C}$ and $\mathbf{K}$ are the mass, damping and stiffness matrices, respectively. Depending on the problem, $n$ is typically  between $10^2$ and $10^5$.
 
 	
 \vspace{0.2cm}
@@ -83,54 +83,49 @@ are formulated using the Finite-Element Method (FEM)~\cite{Bathe2014}, where $\m
 %
 When uncertainty is present, the initial displacements and velocities are modeled as $\mathbf{x}(0) \in \mathcal{X}_0$ and $\mathbf{x}'(0) \in \mathcal{V}_0$ respectively. %
 %
-In \cite{forets2021combining} a novel approach for time integration of structural dynamics equations based on set-based techniques was presented. It allows to compute, in a single integration, the solution sets (or \emph{flowpipes}) that include all exact trajectories under uncertainties in the initial conditions and applied loads.
+In \cite{forets2021combining} a novel approach for time integration of solid dynamics equations based on set-based techniques was presented. %
 %
-Such solution sets cannot be obtained with standard numerical integrators because they are designed to propagate initial points, not sets.
+The approach allows to compute, in a single integration, the solution sets (or \emph{flowpipes}) that include all exact trajectories under uncertainties in the initial conditions and applied loads.
+%
+Such solution sets cannot be obtained using standard numerical integrators, since they are designed to propagate initial points, not sets.
 
 \vspace{0.2cm}
 
 
 \emph{Set-Based FEM Implementation and Application.} %
 %
[…]
