+++ paper/paper.tex
@@ -54,7 +54,7 @@
 
 \emph{The Set-Based Approach.} %
 %
-Real-world problems usually require the resolution of ODEs with uncertainties in the initial conditions or in the constant term. %
+Many real-world problems require the resolution of ODEs with uncertainties in initial conditions or in applied loads
 %
 Obtaining solutions considering these uncertainties is a challenging task, particularly in large scale systems. %
 %
@@ -68,7 +68,7 @@ The set-based approach consists in the construction of sets that contain all the
 %
 In problems such as wave propagation or structural vibrations, solid dynamics problems are formulated. %
 %
-In these cases, large ODEs systems of the form:
+In these cases, large systems of ODEs of the form:
 %
 \begin{equation}\label{eq:second_order}
 \mathbf{M} \mathbf{x}''(t) + \mathbf{C}\mathbf{x}'(t) + \mathbf{K}\mathbf{x}(t) = \mathbf{F}(t),
