### 2022-01-22 #95 @jorgepz
@carstenbauer we updated the article taking in consideration all the contributions and comments made by the reviewers @lmiq and @PetrKryslUCSD . Issues:
 - https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/issues/16
 - https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/issues/15
 - https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/issues/12
 - https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/issues/11
 - https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/issues/6
 - https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/issues/1
