### 2022-01-27 #95 @mforets
> We allow an extension (beyond the 1 page limit) as part of addressing the comments by the reviewers.

To clarify, instead of adding a new example, we attempted to clarify the context of the existing one for people that are not familiar with the topic.
