### 2022-01-27 #95 @mforets
Maybe we should also add a note that there was a whole minisimposium in JuliaCon 2021 about the underlying subject (recording [here](https://www.youtube.com/watch?v=4xni8HBmalE)), with an excellent introductory/motivational talk by Prof. Frehse.
