### 2022-01-31 #95 @lmiq
Minor corrections:

The dot is missing before "Obtaining..." (first paragraph)

The meaning of `\nu_0` and `\xi_0` should be made explicit in "When uncertainty is present... the initial displacements..."
