+++ paper/paper.tex
@@ -54,9 +54,7 @@
 
 \emph{The Set-Based Approach.} %
 %
-Many real-world problems require the resolution of ODEs with uncertainties in initial conditions or in applied loads
-%
-Obtaining solutions considering these uncertainties is a challenging task, particularly in large scale systems. %
+Many real-world problems require the resolution of ODEs with uncertainties in initial conditions or in applied loads. Obtaining solutions considering these uncertainties is a challenging task, particularly in large scale systems. %
 %
 The set-based approach consists in the construction of sets that contain all the feasible solutions of the ODEs~\cite{althoff2020set}. %
 % ---------------
@@ -81,7 +79,7 @@ are formulated using the Finite-Element Method (FEM)~\cite{Bathe2014}, where $\m
 
 \emph{Set-Based Solid Dynamics.} %
 %
-When uncertainty is present, the initial displacements and velocities are modeled as $\mathbf{x}(0) \in \mathcal{X}_0$ and $\mathbf{x}'(0) \in \mathcal{V}_0$ respectively. %
+When uncertainty is present, the initial displacements $\mathbf{x}(0)$ and the initial velocities velocities $\mathbf{x}'(0)$ can be considered as part of corresponding feasible states sets $\mathcal{X}_0$ and  $\mathcal{V}_0$, respectively. %
 %
 In \cite{forets2021combining} a novel approach for time integration of solid dynamics equations based on set-based techniques was presented. %
 %
