### 2022-02-20 c21e93e
Update paper/paper.tex
Co-authored-by: Marcelo Forets <mforets@gmail.com>
