+++ paper/paper.tex
@@ -79,7 +79,7 @@ are formulated using the Finite-Element Method (FEM)~\cite{Bathe2014}, where $\m
 
 \emph{Set-Based Solid Dynamics.} %
 %
-When uncertainty is present, the initial displacements $\mathbf{x}(0)$ and the initial velocities velocities $\mathbf{x}'(0)$ can be considered as part of corresponding feasible states sets $\mathcal{X}_0$ and  $\mathcal{V}_0$, respectively. %
+When uncertainty is present, the initial displacements $\mathbf{x}(0)$ and the initial velocities $\mathbf{x}'(0)$ belong to the feasible sets $\mathcal{X}_0$ and  $\mathcal{V}_0$, respectively. %
 %
 In \cite{forets2021combining} a novel approach for time integration of solid dynamics equations based on set-based techniques was presented. %
 %
