### 2022-02-20 #95 @lmiq
Yes, I'm finished. And I recommend the paper for publication.

I wasn't able to fill a form that was originally requested. I'm not sure if that really matters.
