### 2022-05-03 #95 @carstenbauer
Hey @jorgepz,

Thank you for the reminder. Here a few final stylistic comments.

**Comments**
* Second paragraph
  * You have "problems" twice in one sentence.
  * Please drop the colon before Eq. (1) and the trailing comma in Eq. (1)
* It's suboptimal that the code block and the "Perspectives" are so much disconnected from the rest of the text. It's very easy to mistake the "Perspectives" paragraph as a caption of the code block. Can you please make another attempt to improve this? Some ideas:
  * make the code block a "figure" (e.g. with a label "Alg. 1" and a caption), move the "Perspectives" paragraph to the left column and all figures to the right column. Note that the code block should then appear before Fig. 2 since you reference it before Fig. 2 in the main text.
  * or, alternatively, try exchanging the position of Fig. 2 and code block + "Perspectives" paragraph

Once these points have been addressed we can proceed with the publication of the manuscript!
