+++ paper/ref.bib
@@ -41,7 +41,7 @@
 	pages = {106699},
 	year = {2022},
 	issn = {0045-7949},
-	doi = {https://doi.org/10.1016/j.compstruc.2021.106699},
+	doi = {10.1016/j.compstruc.2021.106699},
 	url = {https://www.sciencedirect.com/science/article/pii/S0045794921002212},
 	author = {Marcelo Forets and Daniel {Freire Caporale} and Jorge M. {Pérez Zerpa}}
 }
@@ -73,7 +73,7 @@
 	title = {{ONSAS.m: an Open Nonlinear Structural Analysis Solver for GNU-Octave/Matlab}},
 	howpublished = {\url{https://github.com/ONSAS/ONSAS.m/}},
 	version = {0.2.4},
-	doi={https://doi.org/10.5281/zenodo.5885614}
+	doi={10.5281/zenodo.5885614}
 }
 
 
