### 2022-05-03 bb2e2db
Create paper.tex
