### 2022-05-03 #95 @carstenbauer
@jorgepz Is it possible to align the top of the code box with the top of the left column?
