### 2022-05-03 #95 @jorgepz
Hi @carstenbauer I believe that anything is _possible_ in _LaTeX_ 🤓
It is difficult to eliminate the above space that lstlisting uses for the caption. I tried including a caption and as you can see here https://github.com/JuliaReach/SetPropagation-FEM-JuliaCon21/blob/cf8eaffe4db9a53ddef66e9c37ac93fbef473eac/paper/paper.pdf the size of the caption in the juliacon class is not changed to footnotesize as it is the case of the figure captions... I assume that that would be a non-desirable style detail... let me know what would you prefer for this. We could also propose a change for the juliacon class inclusing a resize of the lstlistings too...
